add WIP editor for Snowflake public authentication strategy and H2 local datasource specification
